Fix game state not being initialised correctly when restarting
The event bus was retaining references to disposed objects.
So it needed to be lifted up and initialised explicitly.
This allows the game state to be reset when the player reaches a
game over screen and hits reload.

This was a messy commit and includes some other project config changes
as well. Primarily moving back to fractional scaling from integer.
I may want to revisit that in the future if truly seeking a "pixel perfect"
aestehtic, but for now, it seems fine.